Extract closure exporter
